ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
openai-whisper==20231117
scipy==1.11.4
python-dotenv==1.0.0
gunicorn==21.2.0
